+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v. 0.6.0
+ * retritten full iteration for table documents
+ * fixes in yaml dump
+ * additional functions for loading documents from JSON/YAML
+
 ## v. 0.5.0
  * added "extra" field to PiiEntity
  * added add_process_stage() method to PiiEntity